requirements.txt
unstructured
tiktoken
pinecone-client
pypdf
openai
langchain
pandas 
numpy
python-dotenv
sentence_transformers
streamlit
langchain_pinecone
langchain-community
langchain_text_splitters
langchain_openai